Revert to 表示修正
