+++ package.json
@@ -9,12 +9,8 @@
     "preview": "vite preview"
   },
   "devDependencies": {
-    "@types/gl-matrix": "^2.4.5",
     "@webgpu/types": "^0.1.52",
     "typescript": "~5.6.2",
     "vite": "^6.0.5"
-  },
-  "dependencies": {
-    "gl-matrix": "^2.8.1"
   }
 }
