+++ src/main.ts
@@ -1,97 +1,81 @@
-import { initialize } from './initializer';
-import { drawCallback } from './frame';
+import vertexShader from './shader/vertex.wgsl?raw'
+import fragmentShader from './shader/fragment.wgsl?raw'
+
+function frame(device: GPUDevice, context: GPUCanvasContext, pipeline: GPURenderPipeline) {
+    const commandEncoder = device.createCommandEncoder();
+    const textureView = context.getCurrentTexture().createView();
+    const renderPassDescriptor: GPURenderPassDescriptor = {
+        colorAttachments: [
+            {
+                view: textureView,
+                clearValue: { r: 0.0, g: 0.0, b: 0.0, a: 1.0 },
+                loadOp: 'clear',
+                storeOp: 'store',
+            },
+        ],
+    };
+    const passEncoder = commandEncoder.beginRenderPass(renderPassDescriptor);
+    passEncoder.setPipeline(pipeline);
+    passEncoder.draw(3, 1, 0, 0);
+    passEncoder.end();
+
+    device.queue.submit([commandEncoder.finish()]);
+}
 
 async function main() {
-
-    const vertexSize = 4 * 8; // Byte size of one vertex.
-    const positionOffset = 4 * 0;
-    const colorOffset = 4 * 4; // Byte offset of cube vertex color attribute.
-    const vertexCount = 36;
-
-    const vertexArray = new Float32Array([
-        // float4 position, float4 color
-        1, -1, 1, 1, 1, 0, 1, 1,
-        -1, -1, 1, 1, 0, 0, 1, 1,
-        -1, -1, -1, 1, 0, 0, 0, 1,
-        1, -1, -1, 1, 1, 0, 0, 1,
-        1, -1, 1, 1, 1, 0, 1, 1,
-        -1, -1, -1, 1, 0, 0, 0, 1,
-
-        1, 1, 1, 1, 1, 1, 1, 1,
-        1, -1, 1, 1, 1, 0, 1, 1,
-        1, -1, -1, 1, 1, 0, 0, 1,
-        1, 1, -1, 1, 1, 1, 0, 1,
-        1, 1, 1, 1, 1, 1, 1, 1,
-        1, -1, -1, 1, 1, 0, 0, 1,
-
-        -1, 1, 1, 1, 0, 1, 1, 1,
-        1, 1, 1, 1, 1, 1, 1, 1,
-        1, 1, -1, 1, 1, 1, 0, 1,
-        -1, 1, -1, 1, 0, 1, 0, 1,
-        -1, 1, 1, 1, 0, 1, 1, 1,
-        1, 1, -1, 1, 1, 1, 0, 1,
-
-        -1, -1, 1, 1, 0, 0, 1, 1,
-        -1, 1, 1, 1, 0, 1, 1, 1,
-        -1, 1, -1, 1, 0, 1, 0, 1,
-        -1, -1, -1, 1, 0, 0, 0, 1,
-        -1, -1, 1, 1, 0, 0, 1, 1,
-        -1, 1, -1, 1, 0, 1, 0, 1,
-
-        1, 1, 1, 1, 1, 1, 1, 1,
-        -1, 1, 1, 1, 0, 1, 1, 1,
-        -1, -1, 1, 1, 0, 0, 1, 1,
-        -1, -1, 1, 1, 0, 0, 1, 1,
-        1, -1, 1, 1, 1, 0, 1, 1,
-        1, 1, 1, 1, 1, 1, 1, 1,
-
-        1, -1, -1, 1, 1, 0, 0, 1,
-        -1, -1, -1, 1, 0, 0, 0, 1,
-        -1, 1, -1, 1, 0, 1, 0, 1,
-        1, 1, -1, 1, 1, 1, 0, 1,
-        1, -1, -1, 1, 1, 0, 0, 1,
-        -1, 1, -1, 1, 0, 1, 0, 1,
-    ]);
-
-    const instancePositions = new Float32Array([
-        // x, y
-        -5, -5,
-        -5, 0,
-        -5, 5,
-        0, -5,
-        0, 0,
-        0, 5,
-        5, -5,
-        5, 0,
-        5, 5
-    ]);
-
     const canvas = document.querySelector('canvas');
     if (!canvas) {
         throw new Error();
     }
 
-    // deviceの取得
-    const adapter = await navigator.gpu.requestAdapter();
-    if (!adapter) {
+    // WebGPUコンテキストの取得
+    const context = canvas.getContext('webgpu') as GPUCanvasContext | null;
+    if (!context) {
         throw new Error();
     }
-    const device = await adapter.requestDevice();
 
-    const initResult = await initialize({
-        canvas,
-        device,
-        vertexSize,
-        positionOffset,
-        colorOffset,
-        vertexArray,
-        instancePositions
+    // deviceの取得
+    const g_adapter = await navigator.gpu.requestAdapter();
+    if (!g_adapter) {
+        throw new Error();
+    }
+
+    const device = await g_adapter.requestDevice();
+
+    const presentationFormat = navigator.gpu.getPreferredCanvasFormat();
+
+    context.configure({
+        device: device,
+        format: presentationFormat,
+        alphaMode: 'opaque',
     });
 
-    const drawInput = { device, vertexCount, instancePositions, ...initResult };
-    const callback = drawCallback(drawInput);
+    // create a render pipeline
+    const pipeline = device.createRenderPipeline({
+        layout: 'auto',
+        vertex: {
+            module: device.createShaderModule({
+                code: vertexShader,
+            }),
+            entryPoint: 'main',
+        },
+        fragment: {
+            module: device.createShaderModule({
+                code: fragmentShader,
+            }),
+            entryPoint: 'main',
+            targets: [
+                {
+                    format: presentationFormat, // @location(0) in fragment shader
+                }
+            ]
+        },
+        primitive: {
+            topology: 'triangle-list'
+        },
+    });
 
-    requestAnimationFrame(callback);
+    frame(device, context, pipeline);
 }
 
 main();
