+++ src/shader/fragment.wgsl
@@ -1,4 +1,4 @@
 @fragment
-fn main(@location(0) fragColor: vec4<f32>) -> @location(0) vec4<f32> {
-    return fragColor;
+fn main() -> @location(0) vec4<f32> {
+    return vec4<f32>(1.0, 0.0, 0.0, 1.0);
 }
