+++ src/shader/vertex.wgsl
@@ -1,27 +1,10 @@
-
-struct Uniforms {
-    projectionMatrix: mat4x4<f32>,
-    viewMatrix: mat4x4<f32>,
-    worldMatrix: mat4x4<f32>,
-}
-
-struct VertexOutput {
-    @builtin(position) position: vec4<f32>,
-    @location(0) fragColor: vec4<f32>,
-}
-
-@binding(0) @group(0) var<uniform> uniforms: Uniforms;
-
 @vertex
-fn main(
-    @location(0) position: vec4<f32>,
-    @location(1) color: vec4<f32>,
-    @location(2) pos: vec2<f32>,
-) -> VertexOutput {
+fn main(@builtin(vertex_index) vertex_index: u32) -> @builtin(position) vec4<f32> {
+    var pos = array<vec2<f32>, 3>(
+        vec2<f32>(0.0, 0.5),
+        vec2<f32>(-0.5, -0.5),
+        vec2<f32>(0.5, -0.5),
+    );
 
-    var output: VertexOutput;
-    output.position = uniforms.projectionMatrix * uniforms.viewMatrix * uniforms.worldMatrix * (position + vec4<f32>(pos.x, pos.y, 0, 1));
-    output.fragColor = color;
-
-    return output;
+    return vec4<f32>(pos[vertex_index], 0.0, 1.0);
 }
